Доработка 2
Сделаны пункты: 1, 2, 4, 5, 6, 8, 10, 11, 12, 15, 16, 17, 18

+++ src/main/java/com/javaspring/spring2024/domain/Review.java
@@ -54,5 +54,5 @@ public class Review {
      */
     @Setter
     @Column(name = "rating")
-    private Short rating;
+    private Integer rating;
 }
